refactor: form logic split into components
