+++ package.json
@@ -26,6 +26,7 @@
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
     "globals": "^16.3.0",
+    "loadsh": "^0.0.4",
     "prettier": "^3.6.2",
     "react-hook-form": "^7.62.0",
     "react-router": "^7.8.0",
