+++ src/api/editCompany.ts
@@ -2,7 +2,7 @@ import axios, { AxiosError } from "axios";
 import type { CompanyFormType } from "../schema/company.schema";
 import { BackendError } from "../interfaces/BackendError";
 
-export async function editCompany(id: string, data: Partial<CompanyFormType>) {
+export async function editCompany(id: string, data: Partial<Record<keyof CompanyFormType, CompanyFormType[keyof CompanyFormType]>>) {
   try {
     const response = await axios.patch(
       `${import.meta.env.VITE_API_SERVER_URL}/api/v1.0/companies/${id}`,
