--- src/pages/EditPage/EditPage.module.css
@@ -1,15 +1,0 @@
-.container {
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-
-.create_title {
-  font-size: 2rem;
-  font-weight: 700;
-  color: #1a1a1a;
-  text-align: center;
-}
