+++ src/pages/EditPage/EditPage.tsx
@@ -1,13 +1,14 @@
 import { useEffect, useState } from "react";
 import CreateForm from "../../components/CreateForm/CreateForm";
-import styles from "./EditPage.module.css";
+import styles from "../CreatePage/CreatePage.module.css";
 import type { Company } from "../../interfaces/Company.interface";
 import { getCompany } from "../../api/getCompany";
-import { useParams } from "react-router";
+import { useNavigate, useParams } from "react-router";
 import { BackendError } from "../../interfaces/BackendError";
 
 const EditPage = () => {
   const { id } = useParams();
+  const navigate = useNavigate();
   const [company, setCompany] = useState<Company | null>(null);
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [isError, setIsError] = useState<null | string>(null);
@@ -23,10 +24,11 @@ const EditPage = () => {
           setIsError(err.message);
         }
         setIsError("unknown error");
+        navigate('/')
       }
     };
     handleCompany();
-  }, [id]);
+  }, [id, navigate]);
   if (isError) return <div>Error: {isError}</div>;
   if (isLoading) return <div>Loading...</div>;
   return (
